fix: finish moving files
@ParkerH27 only did part 1 in #96
