--- lib/features/pirate_coins/presentation/pirate_coins.dart
@@ -1,1 +1,0 @@
-export "../pirate_coins_page.dart";
