--- lib/features/stats/presentation/stats.dart
@@ -1,1 +1,0 @@
-export "../stats_page.dart";
