+++ lib/features/stats/presentation/stats_page.dart
@@ -5,7 +5,7 @@ import "package:auto_route/auto_route.dart";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
-import "../../widgets/big_card/big_card.dart";
+import "../../../widgets/big_card/big_card.dart";
 
 /// The page at `/pirate-coins/stats`.
 @RoutePage()
